refactor: config folder refactoring
